add game.ts
