+++ frontend/components/game/play/Play.tsx
@@ -4,7 +4,7 @@ import { useSocketStore } from 'store/game/ClientSocket';
 import { usePlayerNamesStore } from 'store/game/PlayerNames';
 import { usePlayStateStore, PlayState } from 'store/game/PlayState';
 import { GameHeader } from 'components/game/play/GameHeader';
-import { GameSetting } from 'types';
+import { GameSetting } from 'types/game';
 
 type Props = {
   gameSetting: GameSetting;
